refactor: useSearch hook to simplify & reuse for other cruds
